Resolvendo erro das chaves estrangeiro com tipo errado (integer para big integer), e arrumando bug da migrate sessoes que estava com o nome errados das tabelas das chaves estrangeiras

+++ database/migrations/2019_05_31_222539_create_guias_table.php
@@ -15,7 +15,7 @@ class CreateGuiasTable extends Migration
     {
         Schema::create('guias', function (Blueprint $table) {
             $table->bigIncrements('id')->primaryKey();
-            $table->integer('paciente_id')->unsigned()->nullable();
+            $table->bigInteger('paciente_id')->unsigned()->nullable();
             $table->foreign('paciente_id')->references('id')->on('pacientes');
             $table->date('data_emissao');
             $table->text('diagnostico');
